[HER-173] Move components under map directory
