+++ src/cameras/map/center-message.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Grid } from '@material-ui/core'
 import PropTypes from 'prop-types'
 
-import { useMessageProps } from 'cameras/list/hooks/use-message-props'
+import { useMessageProps } from 'cameras/map/sidebar/content/list/hooks/use-message-props'
 
 export const CenterMessage = ({ children }) => <Grid {...useMessageProps()}>
   {children}
