+++ src/cameras/map/error.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { Typography } from '@material-ui/core'
 import HighlightOffIcon from '@material-ui/icons/HighlightOff'
 
-import { CenterMessage } from 'cameras/list/center-message'
+import { CenterMessage } from 'cameras/map/center-message'
 
 export const Error = () => <CenterMessage>
   <HighlightOffIcon style={{ fontSize: 70 }} />
