+++ src/cameras/map/hooks/use-content.js
@@ -1,6 +1,6 @@
 import { Content } from 'cameras/map/content'
-import { Loading } from 'cameras/list/loading'
-import { Error } from 'cameras/list/error'
+import { Loading } from 'cameras/map/loading'
+import { Error } from 'cameras/map/error'
 
 export const useContent = () => ({
   success: Content,
