+++ src/cameras/map/loading.jsx
@@ -1,7 +1,7 @@
 import { CircularProgress, Typography } from '@material-ui/core'
 import React from 'react'
 
-import { CenterMessage } from 'cameras/list/center-message'
+import { CenterMessage } from 'cameras/map/center-message'
 
 export const Loading = () => <CenterMessage>
   <CircularProgress disableShrink />
